Revert "change Mysql to MariaDB"
This reverts commit be6344b480a44845c154cd36a0bed4c909a5563c.

+++ src/main/java/org/c3s/edgo/common/beans/DBFullLastInfoBean.java
@@ -305,13 +305,13 @@ public class DBFullLastInfoBean implements DbBean, Serializable {
 	@DataSource({"quered_events", "queredEvents"})
 	@DataTarget("quered_events")
 	@XMLSimple("queredEvents")
-	private java.lang.Integer queredEvents;
+	private java.lang.Long queredEvents;
 	
-	public java.lang.Integer getQueredEvents() {
+	public java.lang.Long getQueredEvents() {
 		return queredEvents;
 	}
 	
-	public DBFullLastInfoBean setQueredEvents(java.lang.Integer value) {
+	public DBFullLastInfoBean setQueredEvents(java.lang.Long value) {
 		this.queredEvents = value;
 		return this;
 	}
