+++ src/main/java/org/c3s/edgo/common/beans/DBHistoryActivityBean.java
@@ -35,13 +35,13 @@ public class DBHistoryActivityBean implements DbBean, Serializable {
 	@DataSource({"event_date", "eventDate"})
 	@DataTarget("event_date")
 	@XMLSimple("eventDate")
-	private java.lang.Integer eventDate;
+	private java.lang.Long eventDate;
 	
-	public java.lang.Integer getEventDate() {
+	public java.lang.Long getEventDate() {
 		return eventDate;
 	}
 	
-	public DBHistoryActivityBean setEventDate(java.lang.Integer value) {
+	public DBHistoryActivityBean setEventDate(java.lang.Long value) {
 		this.eventDate = value;
 		return this;
 	}
